pyproject.toml
[tool.poetry]
name = "ai-categorization"
version = "0.1.0"
description = "AI-powered spend categorization using Anthropic's Claude"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.8"
pandas = "^2.0.0"
anthropic = "^0.49.0"
python-dotenv = "^1.0.0"
tqdm = "^4.67.1"
scikit-learn = "^1.4.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
black = "^24.2.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
categorize = "src.anthropic_utils_dev:main"
describe-vendors = "src.vendor_descriptions:main" 